Execute SDT instructions

See merge request lab2425_summer/armv8_57!26

+++ emulator/src/memory.c
@@ -3,7 +3,7 @@
 
 #define NOT_NULL(ptr) { if (!ptr) LOG_FATAL("NULL pointer"); }
 
-uint8_t read_mem8(processorState_t* state, uint32_t address) {
+uint8_t read_mem8(processorState_t* state, uint64_t address) {
     NOT_NULL(state);
     if (address >= MEMORY_SIZE) {
         LOG_FATAL("Attempt to access outside of memory");
@@ -14,7 +14,7 @@ uint8_t read_mem8(processorState_t* state, uint32_t address) {
 
 // As x86 is little endian this is fine
 
-uint16_t read_mem16(processorState_t* state, uint32_t address) {
+uint16_t read_mem16(processorState_t* state, uint64_t address) {
     NOT_NULL(state);
     if (address + 1 >= MEMORY_SIZE) {
         LOG_FATAL("Attempt to access outside of memory");
@@ -22,7 +22,7 @@ uint16_t read_mem16(processorState_t* state, uint32_t address) {
     return *(uint16_t*)&state->memory[address];
 }
 
-uint32_t read_mem32(processorState_t* state, uint32_t address) {
+uint32_t read_mem32(processorState_t* state, uint64_t address) {
     NOT_NULL(state);
     if (address + 3 >= MEMORY_SIZE) {
         LOG_FATAL("Attempt to access outside of memory");
@@ -30,7 +30,7 @@ uint32_t read_mem32(processorState_t* state, uint32_t address) {
     return *(uint32_t*)&state->memory[address];
 }
 
-uint64_t read_mem64(processorState_t* state, uint32_t address) {
+uint64_t read_mem64(processorState_t* state, uint64_t address) {
     NOT_NULL(state);
     if (address + 7 >= MEMORY_SIZE) {
         LOG_FATAL("Attempt to cacess outside of memory");
@@ -38,7 +38,7 @@ uint64_t read_mem64(processorState_t* state, uint32_t address) {
     return *(uint64_t*)&state->memory[address];
 }
 
-void write_mem8(processorState_t* state, uint32_t address, uint8_t value) {
+void write_mem8(processorState_t* state, uint64_t address, uint8_t value) {
     NOT_NULL(state);
     if (address >= MEMORY_SIZE) {
         LOG_FATAL("Attempt to write outside of memory");
@@ -46,7 +46,7 @@ void write_mem8(processorState_t* state, uint32_t address, uint8_t value) {
     state->memory[address] = value;
 }
 
-void write_mem16(processorState_t* state, uint32_t address, uint16_t value) {
+void write_mem16(processorState_t* state, uint64_t address, uint16_t value) {
     NOT_NULL(state);
     if (address >= MEMORY_SIZE + 1) {
         LOG_FATAL("Attempt to write outside of memory");
@@ -55,7 +55,7 @@ void write_mem16(processorState_t* state, uint32_t address, uint16_t value) {
     *(uint16_t*)&state->memory[address] = value;
 }
 
-void write_mem32(processorState_t* state, uint32_t address, uint32_t value) {
+void write_mem32(processorState_t* state, uint64_t address, uint32_t value) {
     NOT_NULL(state);
     if (address >= MEMORY_SIZE + 3) {
         LOG_FATAL("Attempt to write outside of memory");
@@ -63,7 +63,7 @@ void write_mem32(processorState_t* state, uint32_t address, uint32_t value) {
     *(uint32_t*)&state->memory[address] = value;
 }
 
-void write_mem64(processorState_t* state, uint32_t address, uint64_t value) {
+void write_mem64(processorState_t* state, uint64_t address, uint64_t value) {
     NOT_NULL(state);
     if (address >= MEMORY_SIZE + 7) {
         LOG_FATAL("Attempt to write outside of memory");
